requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.22
pydantic>=1.10
pytest>=7.0
openai>=0.27.0
python-dotenv>=1.0.0
requests>=2.28
scikit-learn>=1.2
